move tests in sub packages
