+++ tests/download/data/test_nprgeneral.py
@@ -2,7 +2,7 @@ import pytest
 
 from pystiche_papers import data
 
-from .asserts import assert_image_downloads_correctly, assert_image_is_downloadable
+from ..asserts import assert_image_downloads_correctly, assert_image_is_downloadable
 
 
 @pytest.mark.slow
