+++ tests/download/paper/test_gatys_ecker_bethge_2016.py
@@ -1,8 +1,8 @@
 import pytest
 
-import pystiche_papers.johnson_alahi_li_2016 as paper
+import pystiche_papers.gatys_ecker_bethge_2016 as paper
 
-from .asserts import assert_image_downloads_correctly, assert_image_is_downloadable
+from ..asserts import assert_image_downloads_correctly, assert_image_is_downloadable
 
 
 @pytest.mark.slow
