+++ tests/download/paper/test_gatys_et_al_2017.py
@@ -2,7 +2,7 @@ import pytest
 
 import pystiche_papers.gatys_et_al_2017 as paper
 
-from .asserts import assert_image_downloads_correctly, assert_image_is_downloadable
+from ..asserts import assert_image_downloads_correctly, assert_image_is_downloadable
 
 
 @pytest.mark.slow
