+++ tests/download/paper/test_ulyanov_et_al_2016.py
@@ -1,8 +1,8 @@
 import pytest
 
-import pystiche_papers.li_wand_2016 as paper
+import pystiche_papers.ulyanov_et_al_2016 as paper
 
-from .asserts import assert_image_downloads_correctly, assert_image_is_downloadable
+from ..asserts import assert_image_downloads_correctly, assert_image_is_downloadable
 
 
 @pytest.mark.slow
